Fix configparser import error
